Added abstract connection class and unified code
