+++ src/connection/AConnection.php
@@ -1,0 +1,59 @@
+<?php
+
+namespace Bolt\connection;
+
+/**
+ * Class AConnection
+ *
+ * @author Michal Stefanak
+ * @link https://github.com/stefanak-michal/Bolt
+ * @package Bolt\connection
+ */
+abstract class AConnection implements IConnection
+{
+
+    /**
+     * @var string
+     */
+    protected $ip;
+
+    /**
+     * @var int
+     */
+    protected $port;
+
+    /**
+     * @var int
+     */
+    protected $timeout;
+
+    /**
+     * AConnection constructor.
+     * @param string $ip
+     * @param int $port
+     * @param int $timeout
+     */
+    public function __construct(string $ip = '127.0.0.1', int $port = 7687, int $timeout = 15)
+    {
+        $this->ip = $ip;
+        $this->port = $port;
+        $this->timeout = $timeout;
+    }
+
+    /**
+     * Print buffer as HEX
+     * @param string $str
+     * @param bool $write
+     */
+    protected function printHex(string $str, bool $write = true)
+    {
+        $str = implode(unpack('H*', $str));
+        echo '<pre>';
+        echo $write ? '> ' : '< ';
+        foreach (str_split($str, 8) as $chunk) {
+            echo implode(' ', str_split($chunk, 2));
+            echo '    ';
+        }
+        echo '</pre>';
+    }
+}
